# backend/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: yourfriends@flexiwan.com
    name: API support
    url: https://flexiwan.com
  description: This is the REST API for flexiWAN management.
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.en.html
  termsOfService: https://flexiwan.com/legal/terms-of-service/
  title: FlexiWAN API
  version: 1.0.0
servers:
- description: Main server
  url: https://manage.flexiwan.com:443/api
paths:
...
  /devices/{id}/routes:
    get:
      operationId: devicesIdRoutesGET
      parameters:
      - description: Numeric ID of the Device to fetch information about
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: The number of items to skip before starting to collect the result
          set
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
      - description: The numbers of items to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceRouteInformation'
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Retrieve device routes information
      tags:
      - Routes
      x-swagger-router-controller: Devices
      x-openapi-router-controller: DevicesController
      x-openapi-router-service: DevicesService
...
  /members:
    get:
      operationId: membersGET
      parameters:
      - description: The number of items to skip before starting to collect the result
          set
        explode: true
        in: query
        name: offset
        required: false
        schema:
          type: integer
        style: form
      - description: The numbers of items to return
        explode: true
        in: query
        name: limit
        required: false
        schema:
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Member'
                type: array
                x-content-type: application/json
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - bearerAuth: []
      - apiKey: []
      summary: Get all Members
      tags:
      - Members
      x-swagger-router-controller: Members
      x-openapi-router-controller: MembersController
      x-openapi-router-service: MembersService
    post:
      operationId: membersPOST
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
          description: Member has been added
          headers:
            Location:
              description: Location of a new resource created
              explode: false
              required: true
              schema:
                type: string
              style: simple
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - bearerAuth: []
      - apiKey: []
      summary: Create new member
      tags:
      - Members
      x-swagger-router-controller: Members
      x-openapi-router-controller: MembersController
      x-openapi-router-service: MembersService
  /members/{id}:
    get:
      operationId: membersIdGET
      parameters:
      - description: Numeric ID of the Member to retrieve
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Member'
                type: array
                x-content-type: application/json
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - bearerAuth: []
      - apiKey: []
      summary: Get Member
      tags:
      - Members
      x-swagger-router-controller: Members
      x-openapi-router-controller: MembersController
      x-openapi-router-service: MembersService
    put:
      operationId: membersIdPUT
      parameters:
      - description: Numeric ID of the account to modify
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
          description: Successful operation
        "400":
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Modify member
      tags:
      - Members
      x-swagger-router-controller: Members
      x-openapi-router-controller: MembersController
      x-openapi-router-service: MembersService
    delete:
      operationId: membersIdDELETE
      parameters:
      - description: Numeric ID of the Member to delete
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Successful operation
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Internal server error
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected error
      security:
      - apiKey: []
      - bearerAuth: []
      summary: Delete member
      tags:
      - Members
      x-swagger-router-controller: Members
      x-openapi-router-controller: MembersController
      x-openapi-router-service: MembersService
...
components:
...
  schemas:
...
    Member:
      example:
        user_email: user_email
        role: owner
        user__id: user__id
        user_name: user_name
        account_name: account_name
        _id: _id
        to: to
        organization_name: organization_name
        account__id: account__id
        group: group
        organization__id: organization__id
      properties:
        _id:
          maxLength: 24
          minLength: 24
          type: string
        user__id:
          maxLength: 24
          minLength: 24
          type: string
        user_name:
          type: string
        user_email:
          type: string
        to:
          type: string
        account_name:
          type: string
        account__id:
          maxLength: 24
          minLength: 24
          type: string
        group:
          type: string
          nullable: true
        organization_name:
          type: string
          nullable: true
        organization__id:
          maxLength: 24
          minLength: 24
          type: string
          nullable: true
        role:
          $ref: '#/components/schemas/Role'
      type: object
...
    DeviceResponseStatus:
      enum:
      - timeout
      - null
      type: string
      nullable: true
    DeviceConnectionStatus:
      enum:
      - connected
      - disconnected
      - unknown
      type: string
...
    DeviceRouteInformation:
      example:
        osRoutes:
        - metric: 0
          destination: destination
          flags: flags
          interface: interface
          gateway: gateway
          mask: mask
        - metric: 0
          destination: destination
          flags: flags
          interface: interface
          gateway: gateway
          mask: mask
        vppRoutes:
        - metric: 6
        - metric: 6
      properties:
        error:
          $ref: '#/components/schemas/DeviceResponseStatus'
        deviceStatus:
          $ref: '#/components/schemas/DeviceConnectionStatus'
        osRoutes:
          items:
            $ref: '#/components/schemas/DeviceRouteOs'
          type: array
        vppRoutes:
          items:
            $ref: '#/components/schemas/DeviceRouteVpp'
          type: array
      type: object
    DeviceRouteOs:
      example:
        metric: 0
        destination: destination
        protocol: static
        interface: interface
        gateway: gateway
      properties:
        metric:
          type: string
        destination:
          type: string
        protocol:
          type: string
        interface:
          type: string
        gateway:
          type: string
      type: object
...
    Error:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
...
  securitySchemes:
    apiKey:
      in: header
      name: APIKEY
      type: apiKey
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http